First pass at support for the Goldy3 .mat data file format.

+++ lib/nav/data/sentera.py
@@ -21,14 +21,16 @@ def isFloat(string):
         return False
     
 def load(flight_dir):
-    imu_data = []
+    result = {}
+    
     gps_data = []
     filter_data = []
 
     # load imu/gps data files
     imu_file = flight_dir + "/imu.csv"
     gps_file = flight_dir + "/gps.csv"
-
+    filter_post = flight_dir + "/filter-post-ins.txt"
+    
     # calibration by plotting and eye-balling (just finding center point, no
     # normalization cooked into calibration.)
     #hx_coeffs = np.array([ 1.0,  -1.5], dtype=np.float64)
@@ -70,12 +72,12 @@ def load(flight_dir):
     # )
 
     # Phantom 3 - Aug 2016 (ellipse cal)
-    mag_affine = np.array(
-        [[ 0.0189725067,  0.0000203615,  0.0002139272, -0.0134053645],
-         [ 0.0000760692,  0.0180178765,  0.0000389461, -1.044762755 ],
-         [ 0.0002417847,  0.0000458039,  0.0171450614,  2.647911793 ],
-         [ 0.          ,  0.          ,  0.          ,  1.          ]]
-    )
+    # mag_affine = np.array(
+    #     [[ 0.0189725067,  0.0000203615,  0.0002139272, -0.0134053645],
+    #      [ 0.0000760692,  0.0180178765,  0.0000389461, -1.044762755 ],
+    #      [ 0.0002417847,  0.0000458039,  0.0171450614,  2.647911793 ],
+    #      [ 0.          ,  0.          ,  0.          ,  1.          ]]
+    # )
     # Phantom 3 - Aug 2016 (ekf cal)
     # mag_affine = np.array(
     #     [[ 0.0181297161,  0.000774339,  -0.002037224 , -0.2576406372],
@@ -83,8 +85,17 @@ def load(flight_dir):
     #      [ 0.0000145964,  0.000267444,   0.0159433791,  2.5630653789],
     #      [ 0.          ,  0.         ,   0.          ,  1.          ]]
     # )
+
+    # 2017-06-07_23-43-50
+    mag_affine = np.array(
+        [[ 0.0167446606, -0.0005639279, -0.0027363919,  0.6582460229],
+         [ 0.0002373517,  0.0175768837,  0.0029620308, -0.1919029845],
+         [ 0.0001372446,  0.0019505221,  0.006486141 ,  1.3921504914],
+         [ 0.          ,  0.          ,  0.          ,  1.          ]]
+    )
     print mag_affine
-    
+
+    result['imu'] = []
     fimu = fileinput.input(imu_file)
     for line in fimu:
         #print line
@@ -104,7 +115,7 @@ def load(flight_dir):
                 hx = -float(hx)
                 hy =  float(hy)
                 hz = -float(hz)
-                mag_orientation = 'newer'
+                mag_orientation = 'random1'
                 if mag_orientation == 'older':
                     #hx_new = hx_func(float(hx))
                     #hy_new = hy_func(float(hy))
@@ -116,6 +127,12 @@ def load(flight_dir):
                     #hy_new = hy_func(float(-hx))
                     #hz_new = hz_func(float(-hz))
                     s = [-hy, -hx, -hz]
+                elif mag_orientation == 'random1':
+                    # remap for 2016-05-12 (0004) data set
+                    #hx_new = hx_func(float(-hy))
+                    #hy_new = hy_func(float(-hx))
+                    #hz_new = hz_func(float(-hz))
+                    s = [-hy, -hx, hz]
                 # mag calibration mapping via mag_affine matrix
                 hs = np.hstack( [s, 1.0] )
                 hf = np.dot(mag_affine, hs)
@@ -133,13 +150,20 @@ def load(flight_dir):
                 imu.ax = ax
                 imu.ay = ay
                 imu.az = az
-                imu.hx = hf[0]
-                imu.hy = hf[1]
-                imu.hz = hf[2]
+                do_mag_transform = False
+                if do_mag_transform:
+                    imu.hx = hf[0]
+                    imu.hy = hf[1]
+                    imu.hz = hf[2]
+                else:
+                    imu.hx = s[0]
+                    imu.hy = s[1]
+                    imu.hz = s[2]
                 #float(hf[0]), float(hf[1]), float(hf[2]),
                 imu.temp = float(temp)
-                imu_data.append( imu )
+                result['imu'].append( imu )
 
+    result['gps'] = []
     fgps = fileinput.input(gps_file)
     for line in fgps:
         if not re.search('Timestamp', line):
@@ -176,8 +200,39 @@ def load(flight_dir):
                 gps.vn = ned[0]
                 gps.ve = ned[1]
                 gps.vd = ned[2]
-                gps_data.append(gps)
-    return imu_data, gps_data, filter_data
+                result['gps'].append(gps)
+
+    result['filter'] = []
+    # load filter (post process) records if they exist (for comparison
+    # purposes)
+    if os.path.exists(filter_post):
+        print 'found filter-post-ins.txt, using that for ekf results'
+        result['filter'] = []
+        ffilter = fileinput.input(filter_post)
+        for line in ffilter:
+            tokens = re.split('[,\s]+', line.rstrip())
+            lat = float(tokens[1])
+            lon = float(tokens[2])
+            if abs(lat) > 0.0001 and abs(lon) > 0.0001:
+                filterpt = nav.structs.NAVdata()
+                filterpt.time = float(tokens[0])
+                filterpt.lat = lat*d2r
+                filterpt.lon = lon*d2r
+                filterpt.alt = float(tokens[3])
+                filterpt.vn = float(tokens[4])
+                filterpt.ve = float(tokens[5])
+                filterpt.vd = float(tokens[6])
+                filterpt.phi = float(tokens[7])*d2r
+                filterpt.the = float(tokens[8])*d2r
+                psi = float(tokens[9])
+                if psi > 180.0:
+                    psi = psi - 360.0
+                if psi < -180.0:
+                    psi = psi + 360.0
+                filterpt.psi = psi*d2r
+                result['filter'].append(filterpt)
+
+    return result
 
 def save_filter_result(filename, data_store):
     f = open(filename, 'w')
@@ -185,10 +240,10 @@ def save_filter_result(filename, data_store):
     for i in range(size):
         line = "%.3f,%.10f,%.10f,%.2f,%.4f,%.4f,%.4f,%.2f,%.2f,%.2f,0" % \
                (data_store.time[i],
-                data_store.nav_lat[i]*180.0/math.pi,
-                data_store.nav_lon[i]*180.0/math.pi,
-                data_store.nav_alt[i], data_store.nav_vn[i],
-                data_store.nav_ve[i], data_store.nav_vd[i],
+                data_store.lat[i]*180.0/math.pi,
+                data_store.lon[i]*180.0/math.pi,
+                data_store.alt[i], data_store.vn[i],
+                data_store.ve[i], data_store.vd[i],
                 data_store.phi[i]*180.0/math.pi,
                 data_store.the[i]*180.0/math.pi,
                 data_store.psi[i]*180.0/math.pi)
@@ -199,6 +254,9 @@ def rewrite_image_metadata_txt(base_dir, data_store):
     meta_file = os.path.join(base_dir, 'image-metadata.txt')
     new_file = os.path.join(base_dir, 'image-metadata-ekf.txt')
 
+    if not os.path.isfile(meta_file):
+        return
+    
     f_out = open(new_file, 'w')
     f_out.write('File Name,Lat (decimal degrees),Lon (decimal degrees),Alt (meters MSL),Yaw (decimal degrees),Pitch (decimal degrees),Roll (decimal degrees),GPS Time (us since epoch)\n')
 
@@ -228,6 +286,9 @@ def rewrite_pix4d_csv(base_dir, data_store):
     meta_file = os.path.join(base_dir, 'image-metadata.txt')
     pix4d_file = os.path.join(base_dir, 'pix4d-ekf.csv')
 
+    if not os.path.isfile(meta_file):
+        return
+    
     f_out = open(pix4d_file, 'w')
     f_out.write('File Name,Lat (decimal degrees),Lon (decimal degrees),Alt (meters MSL),Roll (decimal degrees),Pitch (decimal degrees),Yaw (decimal degrees)\n')
